[WIP][TASK] Initial implementation of full CRUD JSON-API.
When a JSON-API endpoint request is made the following steps are taken:

The routing is setup in a way to support multiple Endpoints, each endpoint can be customised.
When the endpoint is found the InitializeController handles the next steps.

The initializeController accepts the request and handles multiple cases:
 - The endpoint is not configured, exception.
 - The resource is not configured, exception.
 - Resource is not set in AvailableResources, 404 status.

Then the Validation of the request is handled by the ValidatedRequest object which does several checks.
 -

The resolveActionMethodName handles the resolving of the action with the method and the further validation of the request.

Steps taken:
Removed Endpoint Service and replaced it with extensible Adapter implementation.
The Adapter is a extensible layered setup where its possible to use a Doctrine Entity mapping/quering or a custom Model (DTO) like implementation.
The ValidatedRequest class handles the initial Request body where it converts it into a Document object this object can later be parsed and mapped to a Entity or Class.
Simplified the Routes.yaml to always go to index, the "resolveActionMethodName" handles the actionName handling.

Needs:
- Testing custom Adapter
- Getting Relationships
- Putting Relationships
- Index -> Query properly working
- Meta Parameters
 - sort
 - include
 - filter
 - paginate
- Error handling

Later:
- Validator handles validation of the custom request.
- Multiple accessable endpoints